refactor(decompose): split packages/measures/src/health/pressure/pressure-axes.test.ts into folder
