+++ packages/measures/src/health/pressure/pressure-axes.test/repo-root.test.ts
@@ -1,0 +1,3 @@
+import {DEFAULT_REPO_ROOT} from '../../../_shared/discovery/discover-packages'
+
+export const REPO_ROOT = DEFAULT_REPO_ROOT
